Rename model setup to pass R build check
